Push fallback delete to the resilience strategy (#2198)
